refactor: DB schema: decimal prices, supplier contacts
- Change monetary fields from long to decimal for precision
- Change stock/count fields to int for consistency
- Rename OutputInfo.ProductId to ProductVarientId
- Add SupplierContact entity/table for multiple contacts
- Add tax ID to Supplier; SEO fields to Product
- Add CreatedBy/ConfirmedBy/FinishedBy to Input/Output
- Add customer address/phone to Output, parent ref to InputInfo
- Update API contracts and repositories for new types/fields
- Add migration for all schema changes
- Remove unused AuthController.PingTest endpoint

+++ Application/ApiContracts/Input/Requests/UpdateInputInfoRequest.cs
@@ -6,8 +6,8 @@ namespace Application.ApiContracts.Input.Requests
 
         public int? ProductId { get; set; }
 
-        public short? Count { get; set; }
+        public int? Count { get; set; }
 
-        public long? InputPrice { get; set; }
+        public decimal? InputPrice { get; set; }
     }
 }
