+++ Application/Features/Settings/Queries/GetAllSettings/GetAllSettingsQueryHandler.cs
@@ -3,9 +3,9 @@ using MediatR;
 
 namespace Application.Features.Settings.Queries.GetAllSettings;
 
-public sealed class GetAllSettingsQueryHandler(ISettingRepository settingRepository) : IRequestHandler<GetAllSettingsQuery, Dictionary<string, long?>>
+public sealed class GetAllSettingsQueryHandler(ISettingRepository settingRepository) : IRequestHandler<GetAllSettingsQuery, Dictionary<string, decimal?>>
 {
-    public async Task<Dictionary<string, long?>> Handle(
+    public async Task<Dictionary<string, decimal?>> Handle(
         GetAllSettingsQuery request,
         CancellationToken cancellationToken)
     {
